Compiled orthia_lib on linux Compiled orthia_test on linux Working on ELF parsing Added orthia_ui on Linux

+++ src/orthia/orthia_disasm_ui/ui_output_window.cpp
@@ -18,7 +18,7 @@ public:
     {
 
     }
-    void LogData(const ORTHIA_TCHAR* pData, ULONG size)
+    void LogData(const ORTHIA_TCHAR* pData, int size)
     {
         if (auto out = m_outputWindow.lock())
         {
